+++ packages/@webex/internal-plugin-metrics/src/call-diagnostic/call-diagnostic-metrics.ts
@@ -293,11 +293,17 @@ export default class CallDiagnosticMetrics extends StatelessWebexPlugin {
     if (this.webex.internal) {
       // @ts-ignore
       const {device} = this.webex.internal;
+      const {installationId} = device.config || {};
+
       identifiers.userId = device.userId || preLoginId;
       identifiers.deviceId = device.url;
       identifiers.orgId = device.orgId;
       // @ts-ignore
       identifiers.locusUrl = this.webex.internal.services.get('locus');
+
+      if (installationId) {
+        identifiers.machineId = installationId;
+      }
     }
 
     if (meeting?.locusInfo?.fullState) {
@@ -607,6 +613,7 @@ export default class CallDiagnosticMetrics extends StatelessWebexPlugin {
     return this.getErrorPayloadForClientErrorCode({
       clientErrorCode: UNKNOWN_ERROR,
       serviceErrorCode: UNKNOWN_ERROR,
+      payloadOverrides: rawError.payloadOverrides,
       rawErrorMessage,
       httpStatusCode,
     });